feat: pass last and current attribute values to prop effect callback
Note this is the *attribute* value (i.e. a string) not the *property*
value (could be anything). Ideally we want the property value. For
example if we have a boolean property, we want to pass `true` or
`false`, not a string.

Once we have functions that turn property values to and from their
string representation, it should be possible to pass the property values
to the callback.

At the moment since we only support string properties, this is not a
problem.