feat: switch to Idris 2
It turns out Idris 1 can't have a function as a return type from
JavaScript. Since we'll need this for writing effects that use property
getters, I'm switching to Idris 2, which seems to support returning
functions from JavaScript just fine.

This breaks the code completely, but I need to rewrite most of it anyway
so I'll fix it in a later commit.