feat: use functions to build a custom element description
Previously we were building separate custom elements, such as the custom
element with a prop, or the custom element with a side-effect. Now we
build a description of a custom element and in the last step pass the
description to a function that builds the custom element.

This will make it possible to combine the things we want a custom
element to do using do notation (to be implemented later).

All custom element features have been removed except for props, but they
will be re-added later.