fix: Try (and partly fail) to make Depend work with >> and >>=
The part that doesn't work is depending on a Custom Element where the
feature you want is on the right-hand side of >>= e.g.

    Depend (Prop String "hello") (Pure () >>= \_ => Prop String "hello")

The problem is that since the right-hand side of >>= can use the value
from the left-hand side, you can't be sure the feature you want actually
exists. However, if we can show (as in the case above) that regardless
of the value on the left, we will always have the feature we want on the
right (i.e it does not depend on it), then we can say the custom element
has that feature.

That's what I was trying to do but I haven't got it quite working yet.
The broken parts are left in the code commented out.